+++ src/cryo_challenge/_commands/run_map2map_pipeline.py
@@ -6,7 +6,7 @@ import argparse
 import os
 import yaml
 
-from .._map_to_map.map_to_map_distance_matrix import run
+from .._map_to_map.map_to_map_pipeline import run
 from ..data._validation.config_validators import validate_input_config_mtm
