+++ src/cryo_challenge/_map_to_map/map_to_map_pipeline.py
